bugfix: correctly sign attributes
Before this commit all signatures were incorrect, as they only signed
for a message of an empty byte array. Now the bytes of the CID are
actually signed.

Upgrading @noble/ed25519 is not strictly necessary, but it is the first
version that catches this bug, so the upgrade helps validate that the
changes I made to the code are correct.